pulled out shared snapshotdelegate
